main and details page refactoring
