+++ client/src/components/DailyWeatherCard/styles.module.scss
@@ -1,6 +1,6 @@
 .dayWeatherBox {
   text-align: left;
-  padding: 10px ;
+  padding: 10px;
   text-align: center;
   position: relative;
   height: 205px;
