+++ client/src/components/Loader/styles.module.scss
@@ -72,7 +72,6 @@
 }
 
 @media screen and (min-width: 740px) {
-
   .loaderContainer {
     height: calc(100vh - 180px);
     margin-top: 100px;
